+++ src/function/atan2function.cpp
@@ -13,8 +13,8 @@ Value* Atan2Function::evaluate(Context* ctx)
 	NumberValue* yVal=dynamic_cast<NumberValue*>(getParameterArgument(ctx,0));
 	NumberValue* xVal=dynamic_cast<NumberValue*>(getParameterArgument(ctx,1));
 	if(yVal&&xVal) {
-		double y=yVal->getNumber();
-		double x=xVal->getNumber();
+		decimal y=yVal->getNumber();
+		decimal x=xVal->getNumber();
 
 		return new NumberValue(atan2(y,x));
 	}
