+++ src/function/coshfunction.cpp
@@ -11,7 +11,7 @@ Value* CoshFunction::evaluate(Context* ctx)
 {
 	NumberValue* numVal=dynamic_cast<NumberValue*>(getParameterArgument(ctx,0));
 	if(numVal) {
-		double num=numVal->getNumber();
+		decimal num=numVal->getNumber();
 
 		return new NumberValue(cosh(num));
 	}
