+++ src/function/radfunction.cpp
@@ -29,7 +29,7 @@ Value* RadFunction::evaluate(Context* ctx)
 {
 	NumberValue* numVal=dynamic_cast<NumberValue*>(getParameterArgument(ctx,0));
 	if(numVal) {
-		double num=numVal->getNumber();
+		decimal num=numVal->getNumber();
 
 		return new NumberValue(num*M_TAU/360.0);
 	}
