+++ src/function/roundfunction.cpp
@@ -7,7 +7,7 @@ RoundFunction::RoundFunction() : Function("round")
 	addParameter("value");
 }
 
-static double fround(double a)
+static decimal fround(decimal a)
 {
 	return a > 0 ? floor(a+0.5) : ceil(a-0.5);
 }
@@ -16,7 +16,7 @@ Value* RoundFunction::evaluate(Context* ctx)
 {
 	NumberValue* numVal=dynamic_cast<NumberValue*>(getParameterArgument(ctx,0));
 	if(numVal) {
-		double num=numVal->getNumber();
+		decimal num=numVal->getNumber();
 
 		return new NumberValue(fround(num));
 	}
