+++ src/function/signfunction.cpp
@@ -6,7 +6,7 @@ SignFunction::SignFunction() : Function("sign")
 	addParameter("value");
 }
 
-double SignFunction::sign(double num)
+decimal SignFunction::sign(decimal num)
 {
 	return num<0?-1.0:num>0?1.0:0.0;
 }
@@ -15,7 +15,7 @@ Value* SignFunction::evaluate(Context* ctx)
 {
 	NumberValue* numVal=dynamic_cast<NumberValue*>(getParameterArgument(ctx,0));
 	if(numVal) {
-		double num=numVal->getNumber();
+		decimal num=numVal->getNumber();
 
 		return new NumberValue(sign(num));
 	}
