+++ src/function/sqrtfunction.cpp
@@ -29,7 +29,7 @@ Value* SqrtFunction::evaluate(Context* ctx)
 {
 	NumberValue* numVal=dynamic_cast<NumberValue*>(getParameterArgument(ctx,0));
 	if(numVal) {
-		double num=numVal->getNumber();
+		decimal num=numVal->getNumber();
 
 		return new NumberValue(sqrt(num));
 	}
