+++ src/module/polyhedronmodule.cpp
@@ -41,7 +41,7 @@ Node* PolyhedronModule::evaluate(Context* ctx)
 		VectorValue* surface=dynamic_cast<VectorValue*>(s);
 		foreach(Value* indexVal,surface->getChildren()) {
 			NumberValue* indexNum=dynamic_cast<NumberValue*>(indexVal);
-			double index = indexNum->getNumber();
+			decimal index = indexNum->getNumber();
 			VectorValue* point=dynamic_cast<VectorValue*>(children.at(index));
 			Point pt = point->getPoint();
 			p->appendVertex(pt);
