+++ src/module/polylinemodule.cpp
@@ -64,7 +64,7 @@ Node* PolylineModule::evaluate(Context* ctx)
 			p->createPolygon();
 			foreach(Value* indexVal,lineVec->getChildren()) {
 				NumberValue* indexNum=dynamic_cast<NumberValue*>(indexVal);
-				double index = indexNum->getNumber();
+				decimal index = indexNum->getNumber();
 				VectorValue* point=dynamic_cast<VectorValue*>(points.at(index));
 				Point pt = point->getPoint();
 				p->appendVertex(pt);
